implemented apis, added services and packages in catalog

+++ app/src/main/java/life/mibo/android/models/product/Services.kt
@@ -24,7 +24,7 @@ class Services(data: List<Data?>?) : BaseResponse<List<Services.Data?>?>(data) {
         @SerializedName("created_by")
         var createdBy: String?,
         @SerializedName("currency")
-        var currency: Int?,
+        var currency: Double?,
         @SerializedName("currency_type")
         var currencyType: String?,
         @SerializedName("description")
